Refactored Instance model to another package... cleaner like this
