+++ src/main/scala/de/upb/cs/swt/delphi/crawler/instancemanagement/InstanceRegistry.scala
@@ -25,8 +25,8 @@ import akka.http.scaladsl.model._
 import akka.http.scaladsl.unmarshalling.Unmarshal
 import akka.stream.ActorMaterializer
 import de.upb.cs.swt.delphi.crawler.{AppLogging, Configuration, Crawler}
-import de.upb.cs.swt.delphi.crawler.io.swagger.client.model.InstanceEnums.ComponentType
-import de.upb.cs.swt.delphi.crawler.io.swagger.client.model.{Instance, JsonSupport}
+import InstanceEnums.ComponentType
+import de.upb.cs.swt.delphi.crawler.io.swagger.client.model.JsonSupport
 
 import scala.concurrent.duration.Duration
 import scala.concurrent.{Await, ExecutionContext, Future}
